Address @IngridPuppets comments
