+++ services/src/main/java/org/keycloak/protocol/oid4vc/issuance/mappers/OID4VCUserAttributeMapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Red Hat, Inc. and/or its affiliates
+ * Copyright 2025 Red Hat, Inc. and/or its affiliates
  * and other contributors as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,21 +18,38 @@
 package org.keycloak.protocol.oid4vc.issuance.mappers;
 
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 import org.keycloak.models.KeycloakSession;
+import org.keycloak.models.ProtocolMapperModel;
+import org.keycloak.models.UserModel;
+import org.keycloak.models.UserSessionModel;
+import org.keycloak.models.utils.KeycloakModelUtils;
 import org.keycloak.protocol.ProtocolMapper;
+import org.keycloak.protocol.oid4vc.OID4VCLoginProtocolFactory;
+import org.keycloak.protocol.oid4vc.model.VerifiableCredential;
 import org.keycloak.provider.ProviderConfigProperty;
 
 /**
- * Allows adding user attributes to the credential subject
+ * Allows adding user properties to the credential subject
  *
  * @author <a href="https://github.com/wistefan">Stefan Wiedemann</a>
  */
-public class OID4VCUserAttributeMapper extends OID4VCUserPropertyMapper {
+public class OID4VCUserAttributeMapper extends OID4VCMapper {
 
-    public static final String MAPPER_ID = "oid4vc-unmanaged-attribute-mapper";
+    public static final String MAPPER_ID = "oid4vc-user-attribute-mapper";
     public static final String AGGREGATE_ATTRIBUTES_KEY = "aggregateAttributes";
+    public static final List<String> SELECTABLE_PROPERTIES = List.of(UserModel.USERNAME,
+                                                                     UserModel.LOCALE,
+                                                                     UserModel.FIRST_NAME,
+                                                                     UserModel.LAST_NAME,
+                                                                     UserModel.EMAIL,
+                                                                     UserModel.EMAIL_VERIFIED);
 
     private static final List<ProviderConfigProperty> CONFIG_PROPERTIES = new ArrayList<>();
 
@@ -40,16 +57,16 @@ public class OID4VCUserAttributeMapper extends OID4VCUserPropertyMapper {
         ProviderConfigProperty subjectPropertyNameConfig = new ProviderConfigProperty();
         subjectPropertyNameConfig.setName(CLAIM_NAME);
         subjectPropertyNameConfig.setLabel("Claim Name");
-        subjectPropertyNameConfig.setHelpText("The name of the claim added to the credential subject that is extracted " +
-                                                      "from the user attributes.");
+        subjectPropertyNameConfig.setHelpText("The user attributes to be added to the credential subject.");
         subjectPropertyNameConfig.setType(ProviderConfigProperty.STRING_TYPE);
         CONFIG_PROPERTIES.add(subjectPropertyNameConfig);
 
         ProviderConfigProperty userAttributeConfig = new ProviderConfigProperty();
         userAttributeConfig.setName(USER_ATTRIBUTE_KEY);
-        userAttributeConfig.setLabel("User property");
-        userAttributeConfig.setHelpText("The user attributes to be added to the credential subject.");
-        userAttributeConfig.setType(ProviderConfigProperty.STRING_TYPE);
+        userAttributeConfig.setLabel("User attribute");
+        userAttributeConfig.setHelpText("The user attribute to be added to the credential subject.");
+        userAttributeConfig.setType(ProviderConfigProperty.USER_PROFILE_ATTRIBUTE_LIST_TYPE);
+        userAttributeConfig.setOptions(SELECTABLE_PROPERTIES);
         CONFIG_PROPERTIES.add(userAttributeConfig);
 
         ProviderConfigProperty aggregateAttributesConfig = new ProviderConfigProperty();
@@ -65,6 +82,41 @@ public class OID4VCUserAttributeMapper extends OID4VCUserPropertyMapper {
         return CONFIG_PROPERTIES;
     }
 
+    public void setClaim(VerifiableCredential verifiableCredential,
+                         UserSessionModel userSessionModel) {
+        // nothing to do for the mapper.
+    }
+
+    @Override
+    public void setClaim(Map<String, Object> claims, UserSessionModel userSessionModel) {
+        List<String> attributePath = getMetadataAttributePath();
+        String propertyName = attributePath.get(attributePath.size() - 1);
+        String userAttribute = mapperModel.getConfig().get(USER_ATTRIBUTE_KEY);
+        boolean aggregateAttributes = Optional.ofNullable(mapperModel.getConfig().get(AGGREGATE_ATTRIBUTES_KEY))
+                .map(Boolean::parseBoolean).orElse(false);
+        Collection<String> attributes =
+                KeycloakModelUtils.resolveAttribute(userSessionModel.getUser(), userAttribute,
+                        aggregateAttributes);
+        attributes.removeAll(Collections.singleton(null));
+        if (!attributes.isEmpty()) {
+            claims.put(propertyName, String.join(",", attributes));
+        }
+    }
+
+    public static ProtocolMapperModel create(String mapperName, String userAttribute, String propertyName,
+                                             boolean aggregateAttributes) {
+        ProtocolMapperModel mapperModel = new ProtocolMapperModel();
+        mapperModel.setName(mapperName);
+        Map<String, String> configMap = new HashMap<>();
+        configMap.put(CLAIM_NAME, propertyName);
+        configMap.put(USER_ATTRIBUTE_KEY, userAttribute);
+        configMap.put(AGGREGATE_ATTRIBUTES_KEY, Boolean.toString(aggregateAttributes));
+        mapperModel.setConfig(configMap);
+        mapperModel.setProtocol(OID4VCLoginProtocolFactory.PROTOCOL_ID);
+        mapperModel.setProtocolMapper(MAPPER_ID);
+        return mapperModel;
+    }
+
     @Override
     public String getDisplayType() {
         return "User Attribute Mapper";
@@ -72,7 +124,7 @@ public class OID4VCUserAttributeMapper extends OID4VCUserPropertyMapper {
 
     @Override
     public String getHelpText() {
-        return "Maps user attributes to credential claims.";
+        return "Maps user attributes or properties to credential claims.";
     }
 
     @Override
