+++ services/src/main/resources/META-INF/services/org.keycloak.protocol.ProtocolMapper
@@ -51,7 +51,6 @@ org.keycloak.protocol.oidc.mappers.NonceBackwardsCompatibleMapper
 org.keycloak.protocol.oid4vc.issuance.mappers.OID4VCSubjectIdMapper
 org.keycloak.protocol.oid4vc.issuance.mappers.OID4VCStaticClaimMapper
 org.keycloak.protocol.oid4vc.issuance.mappers.OID4VCTargetRoleMapper
-org.keycloak.protocol.oid4vc.issuance.mappers.OID4VCUserPropertyMapper
 org.keycloak.protocol.oid4vc.issuance.mappers.OID4VCUserAttributeMapper
 org.keycloak.protocol.oid4vc.issuance.mappers.OID4VCTypeMapper
 org.keycloak.protocol.oid4vc.issuance.mappers.OID4VCContextMapper
